Update RunSDDC.ps1
Customer gave feedback that when they tried to view the zip this script was downloading, it would not allow them to view it - thus they were concerned to run this against their cluster. Changed invoke-webrequest to absolute path of https://github.com/DellProSupportGse/PrivateCloud.DiagnosticInfo/archive/master.zip to allow those that would like to see it/download it easier before it is extracted and ran on their cluster.